+++ src/library/DIPS.Mobile.UI/Components/Pages/dotnet/ContentPage.navigationbar.cs
@@ -1,0 +1,8 @@
+namespace DIPS.Mobile.UI.Components.Pages;
+
+public partial class ContentPage
+{
+    private partial void ApplyNavigationBarAppearanceOnPlatform()
+    {
+    }
+}